Move server-requests to components
